split apart lh5 classes
